+++ packages/adminui-vue/utils/Form/Form.js
@@ -13,7 +13,7 @@ const {
   transformCollections,
   enrichFieldList
 } = require('./helpers')
-const createValidatorInstance = require('./validation')
+const Validator = require('./validation')
 const UB = require('@unitybase/ub-pub')
 
 /**
@@ -116,7 +116,6 @@ class UForm {
     this.modalClass = modalClass
     this.modalWidth = modalWidth
 
-    this.validatorInstance = undefined
     this.validator = undefined
     this.customValidationMixin = undefined
 
@@ -210,7 +209,7 @@ class UForm {
       instanceID: this.instanceID,
       parentContext: (this.props && this.props.parentContext) ? this.props.parentContext : undefined,
       collections,
-      validator: () => this.validatorInstance && this.validatorInstance.$v,
+      validator: () => this.validator,
       beforeInit: beforeInit ? () => beforeInit.call(this, this.$store) : null,
       inited: inited ? () => inited.call(this, this.$store) : null,
       beforeSave: beforeSave ? () => beforeSave.call(this, this.$store) : null,
@@ -270,14 +269,19 @@ class UForm {
       )
     }
 
+    let $v
+    let getValidationState
+
     if (this.isValidationUsed) {
-      this.validatorInstance = createValidatorInstance(
+      this.validator = new Validator(
         this.$store,
         this.entitySchema,
         this.fieldList,
         this.customValidationMixin
       )
-      this.validator = this.validatorInstance.$v
+
+      $v = this.validator.getValidationState()
+      getValidationState = () => this.validator.getValidationState()
     }
 
     if (this.isProcessingUsed) {
@@ -293,8 +297,8 @@ class UForm {
         component: this.component,
         props: this.props,
         store: this.$store,
-        validator: this.validator,
-        getValidationState: () => this.validatorInstance && this.validatorInstance.$v,
+        $v,
+        getValidationState,
         title: this.title,
         titleTooltip: this.titleTooltip,
         modalClass: this.modalClass,
@@ -320,8 +324,8 @@ class UForm {
         component: this.component,
         props: this.props,
         store: this.$store,
-        validator: this.validator,
-        getValidationState: () => this.validatorInstance && this.validatorInstance.$v,
+        $v,
+        getValidationState,
         title: this.title,
         titleTooltip: this.titleTooltip,
         tabId: this.tabId,
@@ -339,8 +343,8 @@ class UForm {
         component: this.component,
         props: this.props,
         store: this.$store,
-        validator: this.validator,
-        getValidationState: () => this.validatorInstance && this.validatorInstance.$v,
+        $v,
+        getValidationState,
         title: this.title,
         titleTooltip: this.titleTooltip,
         target: this.target,