+++ packages/adminui-vue/utils/Form/validation.js
@@ -1,44 +1,89 @@
-module.exports = createValidatorInstance
+/* global UB */
 
 const Vue = require('vue')
+// vuex required for type checking
+// eslint-disable-next-line no-unused-vars
+const Vuex = require('vuex')
 const { validationMixin } = require('vuelidate')
 const { required } = require('vuelidate/lib/validators/index')
+
 const { mapInstanceFields } = require('./helpers')
 
-/**
- * Track Instance module data and validate it
- * Check entity schema attributes and set required to
- * props which have !allowNull param and defaultView is true
- * @param {Vuex} store Store
- * @param {UBEntity} entitySchema Entity schema
- * @param {Vue} customValidationMixin Custom validation mixin in case when need to override default validation
- * @param {string[]} masterFieldList Field list of master entity
- * @return {object} Vuelidate validation object
- */
-function createValidatorInstance (store, entitySchema, masterFieldList, customValidationMixin = {}) {
-  const requiredFields = entitySchema
-    .filterAttribute(attr => attr.defaultView && !attr.allowNull && masterFieldList.includes(attr.code))
-    .map(a => a.name)
+module.exports = class Validator {
+  /**
+   * Create a Vue instance for form data validation.
+   * Default behavior is to check entity schema attributes with `allowNull=true` and `defaultView=true`.
+   *
+   * `customValidationMixin` can extend default behavior by it own rules.
+   *
+   * @param {Vuex} store Store
+   * @param {UBEntity} entitySchema Entity schema
+   * @param {string[]} masterFieldList Field list of master entity
+   * @param {Vue} [customValidationMixin={}] Custom validations what extends default
+   * @return {Vue} Vue instance
+   */
+  constructor (store, entitySchema, masterFieldList, customValidationMixin = {}) {
+    this._entitySchema = entitySchema
 
-  const defaultValidationMixin = {
-    computed: mapInstanceFields(entitySchema.getAttributeNames()),
+    const requiredFields = entitySchema
+      .filterAttribute(attr => attr.defaultView && !attr.allowNull && masterFieldList.includes(attr.code))
+      .map(a => a.name)
 
-    validations () {
-      return requiredFields.reduce((obj, field) => {
-        obj[field] = { required }
-        return obj
-      }, {})
+    const defaultValidationMixin = {
+      computed: mapInstanceFields(entitySchema.getAttributeNames()),
+
+      validations () {
+        return Object.fromEntries(requiredFields.map(field => [field, { required }]))
+      }
+    }
+
+    this._vueInstance = new Vue({
+      store,
+      mixins: [
+        validationMixin,
+        defaultValidationMixin,
+        customValidationMixin
+      ]
+    })
+  }
+
+  /**
+   * Returns the current state of validation. The method is useful when you have dynamic validation
+   * @returns {object} Vuelidate object
+   */
+  getValidationState () {
+    return this._vueInstance.$v
+  }
+
+  /**
+   * Validates form data with Vuelidate. If validation is failed `UBAbortError` will be thrown
+   * @throws {UB.UBAbortError}
+   */
+  validateForm () {
+    const { $v } = this._vueInstance
+    $v.$touch()
+    if ($v.$error) {
+      const masterEntityName = this._entitySchema.code
+      const invalidFields = Object.keys($v.$params).filter(f => $v[f].$invalid)
+      const errors = invalidFields.map(field => {
+        const formFieldCaption = this._vueInstance[`${field}:caption`]
+        if (formFieldCaption) {
+          return formFieldCaption
+        }
+        const localeString = `${masterEntityName}.${field}`
+        return UB.i18n(localeString) === localeString ? field : UB.i18n(localeString)
+      })
+      const errMsg = UB.i18n('validationError', errors.join(', '))
+      const err = new UB.UBError(errMsg)
+      UB.showErrorWindow(err)
+      throw new UB.UBAbortError(errMsg)
     }
   }
 
-  const instance = new Vue({
-    store,
-    mixins: [
-      validationMixin,
-      defaultValidationMixin,
-      customValidationMixin
-    ]
-  })
-
-  return instance
+  /**
+   * Reset validation state
+   */
+  reset () {
+    this._vueInstance.$v.$reset()
+  }
 }